+++ typescript/sdk/src/index.ts
@@ -779,10 +779,7 @@ export {
   WarpTypedTransaction,
 } from './warp/types.js';
 export { WarpCore, WarpCoreOptions } from './warp/WarpCore.js';
-export {
-  getPendingEvmTimelockControllerTransactions,
-  getPendingExecutableEvmTimelockControllerTransactions,
-} from './providers/transactions/submitter/ethersV5/timelock/transactions.js';
+export { EvmTimelockReader } from './timelock/evm/EvmTimelockReader.js';
 export { EvmTimelockDeployer } from './timelock/evm/EvmTimelockDeployer.js';
 export {
   TimelockConfig,
