+++ typescript/sdk/src/timelock/evm/constants.ts
@@ -1,5 +1,8 @@
 import { keccak256 } from 'ethers/lib/utils.js';
 
+export const ZERO_32_BYTES =
+  '0x0000000000000000000000000000000000000000000000000000000000000000';
+
 export const PROPOSER_ROLE: string = keccak256(Buffer.from('PROPOSER_ROLE'));
 export const EXECUTOR_ROLE: string = keccak256(Buffer.from('EXECUTOR_ROLE'));
 export const CANCELLER_ROLE: string = keccak256(Buffer.from('CANCELLER_ROLE'));
